More stable presentation
Reworked how presentations work - the first track is loaded, and the rest are queued. User controls when the next track plays, which tells the frontend web player to skip and seek to the correct spot.

The difference with this approach, is that any calls to the server or spotify are done at the start of the presentation. All other controls are based on the frontend.